NOISSUE Move name of error property to api module
Other modules may also use this information, and we want it to be consistent across all modules.
