+++ api/src/main/java/energy/eddie/api/agnostic/GlobalConfig.java
@@ -1,0 +1,7 @@
+package energy.eddie.api.agnostic;
+
+public final class GlobalConfig {
+    public static final String ERRORS_PROPERTY_NAME = "errors";
+
+    private GlobalConfig() {}
+}
