+++ region-connectors/shared/src/main/java/energy/eddie/spring/regionconnector/extensions/RegionConnectorsCommonControllerAdvice.java
@@ -25,10 +25,11 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
 
+import static energy.eddie.api.agnostic.GlobalConfig.ERRORS_PROPERTY_NAME;
+
 @RegionConnectorExtension
 @RestControllerAdvice
 public class RegionConnectorsCommonControllerAdvice {
-    public static final String ERRORS_PROPERTY_NAME = "errors";
     public static final String ERRORS_JSON_PATH = "$." + ERRORS_PROPERTY_NAME;
     private static final Logger LOGGER = LoggerFactory.getLogger(RegionConnectorsCommonControllerAdvice.class);
